Time: 0 ms (100%), Space: 7.5 MB (45.4%) - LeetHub

+++ 0400-nth-digit/0400-nth-digit.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int findNthDigit(int n) {
-        long long digits = 1;
+        int digits = 1;
         long long one = 1;
         long long nine = 9;
         while(n > digits*nine){
